+++ package.json
@@ -4,10 +4,11 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon --exec babel-node app.js",
+    "build": "del-cli --force dist && babel ./src/ -d dist",
+    "start": "npm run seed && node ./dist/app.js",
     "dev": "nodemon --exec babel-node user.seed.js app.js",
     "test": "nyc mocha --recursive --exit  --require @babel/register test/**/*.js",
-    "seed": "nodemon --exec babel-node user.seed.js"
+    "seed": "npm run build && node ./dist/seeds/user.seed.js"
   },
   "keywords": [],
   "author": "Christella",
