--- routes/userMessages-routes.js
@@ -1,25 +1,0 @@
-import express from "express";
-import { deleteMessage } from "./../src/controllers/userMessages-controller";
-import messageMiddleware from "../src/middlewares/messagevalidation";
-import {
-  createMessage,
-  getAllMessages,
-  getMessages,
-} from "./../src/controllers/userMessages-controller";
-
-const router = express.Router();
-
-// Create article
-router.post("", messageMiddleware, createMessage);
-
-// Get all articles
-router.get("/", getAllMessages);
-
-// Get article by id
-router.get("/getOne/:id", getMessages);
-
-//Delete article by id
-router.delete("/deletemessage/:id", deleteMessage);
-
-export default router;
-//module.exports = router;
